AppNergizer
This is the prototype version of AppNergizer created for Windows app.

The main scope of AppNergizer app is to spread the awareness to all
users relating towards saving energy of power.

This is just only a prototype, a full version of this app will be deploy
by end of this year. Thank you.